EcommerceTesting_TS_12/tests/Navigation.test.ts
// Mock DOM implementation
class MockElement {
  classList: string[] = [];
  eventListeners: { [key: string]: (() => void)[] } = {};

  addEventListener(event: string, callback: () => void): void {
    if (!this.eventListeners[event]) {
      this.eventListeners[event] = [];
    }
    this.eventListeners[event].push(callback);
  }

  click(): void {
    if (this.eventListeners['click']) {
      this.eventListeners['click'].forEach(callback => callback());
    }
  }

  classList = {
    add: (className: string) => {
      if (!this.classList.includes(className)) {
        this.classList.push(className);
      }
    },
    remove: (className: string) => {
      const index = this.classList.indexOf(className);
      if (index !== -1) {
        this.classList.splice(index, 1);
      }
    },
    contains: (className: string) => {
      return this.classList.includes(className);
    }
  };
}

// Mock document implementation
const mockDocument = {
  elements: {} as { [key: string]: MockElement },
  getElementById: function(id: string): MockElement | null {
    return this.elements[id] || null;
  },
  createElement: function(tagName: string): MockElement {
    return new MockElement();
  }
};

// Global document object
declare global {
  var document: typeof mockDocument;
}
global.document = mockDocument;

// Function to be tested
function initializeNavigation(): void {
  const bar: HTMLElement | null = document.getElementById('bar');
  const close: HTMLElement | null = document.getElementById('close');
  const nav: HTMLElement | null = document.getElementById('navbar');

  if (bar && nav) {
    bar.addEventListener('click', (): void => {
      nav.classList.add('active');
    });
  }

  if (close && nav) {
    close.addEventListener('click', (): void => {
      nav.classList.remove('active');
    });
  }
}

// Test suite
describe('Navigation Initialization', () => {
  beforeEach(() => {
    // Reset mock document before each test
    document.elements = {
      'bar': document.createElement('div'),
      'close': document.createElement('div'),
      'navbar': document.createElement('nav')
    };
  });

  test('adds active class to navbar when bar is clicked', () => {
    initializeNavigation();
    const bar = document.getElementById('bar');
    const nav = document.getElementById('navbar');
    
    bar?.click();
    
    expect(nav?.classList.contains('active')).toBe(true);
  });

  test('removes active class from navbar when close is clicked', () => {
    initializeNavigation();
    const close = document.getElementById('close');
    const nav = document.getElementById('navbar');
    
    nav?.classList.add('active');
    close?.click();
    
    expect(nav?.classList.contains('active')).toBe(false);
  });

  test('does nothing when elements are not found', () => {
    document.elements = {};
    
    expect(() => initializeNavigation()).not.toThrow();
  });
});

// Test runner implementation
function describe(name: string, fn: () => void) {
  console.log(`Test Suite: ${name}`);
  fn();
}

function test(name: string, fn: () => void) {
  console.log(`  Test: ${name}`);
  try {
    fn();
    console.log('    Passed');
  } catch (error) {
    console.error(`    Failed: ${error}`);
  }
}

function beforeEach(fn: () => void) {
  fn();
}

function expect(received: any) {
  return {
    toBe: (expected: any) => {
      if (received !== expected) {
        throw new Error(`Expected ${expected} but received ${received}`);
      }
    },
    not: {
      toThrow: () => {
        try {
          received();
        } catch (error) {
          throw new Error(`Expected function not to throw, but it threw ${error}`);
        }
      }
    }
  };
}

// Run the tests
describe('Navigation Initialization', () => {
  beforeEach(() => {
    document.elements = {
      'bar': document.createElement('div'),
      'close': document.createElement('div'),
      'navbar': document.createElement('nav')
    };
  });

  test('adds active class to navbar when bar is clicked', () => {
    initializeNavigation();
    const bar = document.getElementById('bar');
    const nav = document.getElementById('navbar');
    
    bar?.click();
    
    expect(nav?.classList.contains('active')).toBe(true);
  });

  test('removes active class from navbar when close is clicked', () => {
    initializeNavigation();
    const close = document.getElementById('close');
    const nav = document.getElementById('navbar');
    
    nav?.classList.add('active');
    close?.click();
    
    expect(nav?.classList.contains('active')).toBe(false);
  });

  test('does nothing when elements are not found', () => {
    document.elements = {};
    
    expect(() => initializeNavigation()).not.toThrow();
  });
});

console.log('All tests completed.');